software/telegram_bot/requirements.txt
python-telegram-bot[job-queue]==22.7
anthropic==0.86.0
influxdb-client==1.50.0
resend==2.26.0
httpx==0.27.0
